+++ src/inc/random_play.h
@@ -10,6 +10,7 @@ and some other fast checks.
 
 #include "board.h"
 #include "cfg_board.h"
+#include "move.h"
 #include "types.h"
