+++ src/inc/tactical.h
@@ -223,75 +223,6 @@ bool is_ladder(
     bool is_black
 );
 
-/*
-Return the minimum amount of liberties of groups with stones adjacent to an
-intersection.
-RETURNS minimum number of liberties found, or NONE
-*/
-u16 min_neighbor_libs(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-);
-
-/*
-Return the maximum amount of liberties of groups with stones adjacent to an
-intersection.
-RETURNS maximum number of liberties found, or 0
-*/
-u8 max_neighbor_libs(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-);
-
-/*
-Tests whether a neighbor group of stone type stone has two liberties.
-RETURNS true if neighbor group is put in atari
-*/
-bool puts_neighbor_in_atari(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-);
-
-/*
-Return the maximum number of stones of a group of stones of value stone;
-adjacent to the intersection m.
-RETURNS maximum number of stones of a group, or 0
-*/
-u16 max_neighbor_group_stones(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-);
-
-/*
-Tests whether two groups have exactly the same liberties.
-RETURNS true if the groups have the exact same liberties
-*/
-bool groups_same_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-);
-
-/*
-Tests whether two groups share at least one liberty.
-RETURNS true if the groups share at least one liberty
-*/
-bool groups_share_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-);
-
-/*
-Counts the number of shared liberties between two groups.
-RETURNS number of shared liberties
-*/
-u8 groups_shared_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-);
 
 /*
 Tests whether group g can be attacked and eventually killed by its opponent,
