+++ src/ob_gen/main.c
@@ -9,19 +9,19 @@ Application for the production of Fuego book from SGF game collections.
 #include <stdio.h>
 #include <string.h>
 
+#include "alloc.h"
 #include "board.h"
+#include "constants.h"
+#include "crc32.h"
 #include "engine.h"
 #include "file_io.h"
-#include "crc32.h"
-#include "timem.h"
+#include "hash_table.h"
 #include "opening_book.h"
 #include "randg.h"
 #include "sgf.h"
-#include "hash_table.h"
 #include "state_changes.h"
 #include "stringm.h"
-#include "alloc.h"
-#include "flog.h"
+#include "timem.h"
 
 #define MAX_FILES 500000
 #define TABLE_BUCKETS 4957
@@ -262,7 +262,6 @@ o be used. (default: %u)\n", minimum_turns);
     }
 
     alloc_init();
-    config_logging(DEFAULT_LOG_MODES);
     assert_data_folder_exists();
 
     char * ts = alloc();
