+++ src/tactical.c
@@ -4,8 +4,8 @@ Tactical functions that make use of a cfg_board structure.
 Tactical analysis in Matilda mostly concerns with two things:
 - Eye shape -- eyes, nakade
 - Life and death -- ladders, seki, 1-2 liberty solvers for killing and saving
-groups, connecting groups by kosumi, bamboo joints, etc for the
-purpose of eye counting.
+groups, connecting groups by kosumi, bamboo joints, etc for the purpose of eye
+counting.
 */
 
 #include "matilda.h"
@@ -34,7 +34,7 @@ extern bool border_top[TOTAL_BOARD_SIZ];
 extern bool border_bottom[TOTAL_BOARD_SIZ];
 
 /* number of active bits for every byte combination */
-extern u8 dyn_active_bits[256];
+extern u8 active_bits_in_byte[256];
 
 
 /*
@@ -1032,149 +1032,6 @@ bool is_ladder(
     assert(0);
 }
 
-
-/*
-Return the minimum amount of liberties of groups with stones adjacent to an
-intersection.
-RETURNS minimum number of liberties found, or NONE
-*/
-u16 min_neighbor_libs(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-){
-    assert(is_board_move(m));
-
-    u16 ret = NONE;
-    if(!border_left[m] && cb->p[m + LEFT] == stone)
-        ret = cb->g[m + LEFT]->liberties;
-    if(!border_right[m] && cb->p[m + RIGHT] == stone && cb->g[m +
-        RIGHT]->liberties < ret)
-        ret = cb->g[m + RIGHT]->liberties;
-    if(!border_top[m] && cb->p[m + TOP] == stone && cb->g[m + TOP]->liberties <
-        ret)
-        ret = cb->g[m + TOP]->liberties;
-    if(!border_bottom[m] && cb->p[m + BOTTOM] == stone && cb->g[m +
-        BOTTOM]->liberties < ret)
-        ret = cb->g[m + BOTTOM]->liberties;
-    return ret;
-}
-
-
-/*
-Return the maximum amount of liberties of groups with stones adjacent to an
-intersection.
-RETURNS maximum number of liberties found, or 0
-*/
-u8 max_neighbor_libs(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-){
-    assert(is_board_move(m));
-
-    u8 ret = 0;
-    if(!border_left[m] && cb->p[m + LEFT] == stone)
-        ret = cb->g[m + LEFT]->liberties;
-    if(!border_right[m] && cb->p[m + RIGHT] == stone && cb->g[m +
-        RIGHT]->liberties > ret)
-        ret = cb->g[m + RIGHT]->liberties;
-    if(!border_top[m] && cb->p[m + TOP] == stone && cb->g[m + TOP]->liberties >
-        ret)
-        ret = cb->g[m + TOP]->liberties;
-    if(!border_bottom[m] && cb->p[m + BOTTOM] == stone && cb->g[m +
-        BOTTOM]->liberties > ret)
-        ret = cb->g[m + BOTTOM]->liberties;
-    return ret;
-}
-
-/*
-Tests whether a neighbor group of stone type stone has two liberties.
-RETURNS true if neighbor group is put in atari
-*/
-bool puts_neighbor_in_atari(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-){
-    if(!border_left[m] && cb->p[m + LEFT] == stone && cb->g[m + LEFT]->liberties
-        == 2)
-        return true;
-    if(!border_right[m] && cb->p[m + RIGHT] == stone && cb->g[m +
-        RIGHT]->liberties == 2)
-        return true;
-    if(!border_top[m] && cb->p[m + TOP] == stone && cb->g[m + TOP]->liberties ==
-        2)
-        return true;
-    return (!border_bottom[m] && cb->p[m + BOTTOM] == stone && cb->g[m +
-        BOTTOM]->liberties == 2);
-}
-
-/*
-Return the maximum number of stones of a group of stones of value stone;
-adjacent to the intersection m.
-RETURNS maximum number of stones of a group, or 0
-*/
-u16 max_neighbor_group_stones(
-    const cfg_board * cb,
-    move m,
-    u8 stone
-){
-    assert(is_board_move(m));
-
-    u16 ret = 0;
-    if(!border_left[m] && cb->p[m + LEFT] == stone)
-        ret = cb->g[m + LEFT]->stones.count;
-    if(!border_right[m] && cb->p[m + RIGHT] == stone && cb->g[m +
-        RIGHT]->stones.count > ret)
-        ret = cb->g[m + RIGHT]->stones.count;
-    if(!border_top[m] && cb->p[m + TOP] == stone && cb->g[m +
-        TOP]->stones.count > ret)
-        ret = cb->g[m + TOP]->stones.count;
-    if(!border_bottom[m] && cb->p[m + BOTTOM] == stone && cb->g[m +
-        BOTTOM]->stones.count > ret)
-        ret = cb->g[m + BOTTOM]->stones.count;
-    return ret;
-}
-
-/*
-Tests whether two groups have exactly the same liberties.
-RETURNS true if the groups have the exact same liberties
-*/
-bool groups_same_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-){
-    return memcmp(g1->ls, g2->ls, LIB_BITMAP_SIZ) == 0;
-}
-
-/*
-Tests whether two groups share at least one liberty.
-RETURNS true if the groups share at least one liberty
-*/
-bool groups_share_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-){
-    for(u8 i = 0; i < LIB_BITMAP_SIZ; ++i)
-        if((g1->ls[i] & g2->ls[i]) > 0)
-            return true;
-    return false;
-}
-
-/*
-Counts the number of shared liberties between two groups.
-RETURNS number of shared liberties
-*/
-u8 groups_shared_liberties(
-    const group * restrict g1,
-    const group * restrict g2
-){
-    u8 ret = 0;
-    for(u8 i = 0; i < LIB_BITMAP_SIZ; ++i)
-        ret += dyn_active_bits[g1->ls[i] & g2->ls[i]];
-    return ret;
-}
